project

+++ app/src/main/java/uk/ac/tees/c2704207/currency/data/remote/DataTransfer/CurrencyDataTransfer.kt
@@ -32,6 +32,6 @@ data class CurrencyDataTransfer(
     val SGD: Double,
     val THB: Double,
     val TRY: Double,
-    val USD: Int,
+    val USD: Double,
     val ZAR: Double
 )
